Add files via upload
A Currency Converter is a simple yet powerful software application that allows users to quickly and accurately convert between different currencies. This project aims to create an intuitive and user-friendly currency conversion tool that can be used by travelers, businesses, and individuals to facilitate international financial transactions and manage foreign currency exchange rates.

Conversion Calculator: Users should be able to input the amount they want to convert, and the application should instantly display the converted amount based on the selected currencies and current exchange rates.

This Currency Converter project has the potential to be a valuable tool for individuals and businesses engaging in international finance and can help users make informed decisions about currency exchanges, ultimately simplifying the process of dealing with different currencies in a globalized world.